refactor from javascript funtions to classes
