+++ index.html
@@ -5,7 +5,7 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <link rel="stylesheet" href="styles.css">
-    <script src="awesomeBooks.js" defer></script>
+    <script type="module" src="awesomeBooks.js" defer></script>
     <title>Awesome Books</title>
   </head>
   <body>
